% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellDivide.R
\name{cellDivide}
\alias{cellDivide}
\title{Division of Bacterial Population}
\usage{
cellDivide(P, Rd, maxPsize, thr)
}
\arguments{
\item{P}{The population of bacteria arranged in a two-dimensional matrix of 0 and 1 values, where 0 represents naive loci and 1 represents mutated loci.}

\item{Rd}{ADD LATER}

\item{maxPsize}{The maximum population size that the bacterial population will reach within a day. Defaults to 2000.}

\item{thr}{Bacteria with an overall fitness below this specified threshold do not reproduce to the next generation. Defaults to 0.51.}
}
\description{
Functions to manipulate individual operations of division within a cell population.
}
\examples{
P2 <- cellDivide(P, Rd = 0.2, maxPsize = 500, thr = 0.51)

}
\author{
Ahmad Abdel-Azim \email{agabdel.azim@gmail.com}

Salma Abdel-Azim \email{salma.abdelazim@gmail.com}

Gamal Abdel-Azim \email{gamal.azim@gmail.com}
}
\keyword{SAR}
\keyword{antibiotic}
\keyword{resistance}
\keyword{simulation}
