% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructPopulation.R
\name{constructPopulation}
\alias{constructPopulation}
\title{Construction of Bacterial Population}
\usage{
constructPopulation(Psize, Nl, Ng, startingFitness)
}
\arguments{
\item{Psize}{Starting popultaion size. Defaults to 300.}

\item{Nl}{Number of loci per gene per section. One-dimensional array of length equal to number of sections. Defaults to c(1, 6, 10).}

\item{Ng}{Number of genes per section. One-dimensional array of length equal to number of sections. Defaults to c(6, 10, 12).}

\item{startingFitness}{Starting population fitness, represnted by the proportion of mutated bacterial cells in the population. Defaults to 0.51.}
}
\description{
TO BE ADDED LATER
}
\author{
Ahmad Abdel-Azim \email {agabdel.azim@gmail.com}

Salma Abdel-Azim \email {salma.abdelazim@gmail.com}

Gamal Abdel-Azim \email {gamal.azim@gmail.com}
}
