% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOverallResistance.R
\name{getOverallResistance}
\alias{getOverallResistance}
\alias{getResistance}
\title{Measure Overall Resistance of Bacterial Population}
\usage{
getResistance(gtype, Nl, Ng)
}
\arguments{
\item{gtype}{ADD LATER}

\item{Nl}{Number of loci per gene per section. One-dimensional array of length equal to number of sections. Defaults to c(1, 6, 10).}

\item{Ng}{Number of genes per section. One-dimensional array of length equal to number of sections. Defaults to c(6, 10, 12).}
}
\description{
TO BE ADDED LATER
}
\author{
Ahmad Abdel-Azim \email {agabdel.azim@gmail.com}

Salma Abdel-Azim \email {salma.abdelazim@gmail.com}

Gamal Abdel-Azim \email {gamal.azim@gmail.com}
}
