% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_data.R
\docType{data}
\name{ms_catalog}
\alias{ms_catalog}
\title{Sample catalog dataset}
\format{
A data frame with the following variables:
\describe{
\item{item}{The item ID, corresponding to the items in the \link{ms_usage} dataset.}
\item{name}{A short description of the item.}
\item{category}{The item category.}
}
}
\source{
Microsoft.
}
\usage{
ms_catalog
}
\description{
Dataset of item descriptions from the Microsoft online store.
}
\keyword{datasets}
