% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_data.R
\docType{data}
\name{ms_usage}
\alias{ms_usage}
\title{Sample usage dataset}
\format{
A data frame with the following variables:
\describe{
\item{user}{The user ID.}
\item{item}{The item ID, corresponding to the items in the \link{ms_catalog} dataset.}
\item{time}{The date and time of the transaction, in POSIXct format.}
}
}
\source{
Microsoft.
}
\usage{
ms_usage
}
\description{
Dataset of anonymised transaction records from the Microsoft online store.
}
\keyword{datasets}
