% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate}
\alias{mutate}
\title{Mutate a Bacterial Population}
\usage{
mutate(bCell, Rm)
}
\arguments{
\item{bCell}{One bacterial cell in the population, represented as a row in the population matrix, P.}

\item{Rm}{Rate of mutation in bacterial population. Defaults to 0.001.}
}
\description{
TO BE ADDED LATER
}
\author{
Ahmad Abdel-Azim \email {agabdel.azim@gmail.com}

Salma Abdel-Azim \email {salma.abdelazim@gmail.com}

Gamal Abdel-Azim \email {gamal.azim@gmail.com}
}
