% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDayAvg.R
\name{plotDayAvg}
\alias{plotDayAvg}
\title{Plot Day Averages for Simulated Scenarios}
\usage{
plotDayAvg(scenario, gen.pd, nDay, Nl, Ng, ylim)
}
\arguments{
\item{scenario}{ADD LATER}

\item{gen.pd}{ADD LATER}

\item{nDay}{ADD LATER}

\item{Nl}{Number of loci per gene per section. One-dimensional array of length equal to number of sections.}

\item{Ng}{Number of genes per section. One-dimensional array of length equal to number of sections.}

\item{ylim}{ADD LATER}
}
\value{
ADD LATER
}
\description{
ADD LATER
}
\details{
ADD LATER
}
\examples{
plotDayAvg(scenario = scenarioSimulate, gen.pd = 24, nDay = 20, Nl = c(1, 2, 3), Ng = c(10, 10, 10), ylim = c(0.5, 0.85))
}
\author{
Ahmad Abdel-Azim \email{agabdel.azim@gmail.com}

Salma Abdel-Azim \email{salma.abdelazim@gmail.com}

Gamal Abdel-Azim \email{gamal.azim@gmail.com}
}
\keyword{SAR}
\keyword{antibiotic}
\keyword{resistance}
\keyword{simulation}
