% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSimulation.R
\name{runSimulation}
\alias{runSimulation}
\title{Simulation of Antibiotic Resistance in Bacteria over Multiple Exposures}
\usage{
runSimulation(gen.interval, Rm, Ng, Nl, Psize, startingFitness, thr, nDays, maxPsize)
}
\arguments{
\item{gen.interval}{Generation interval in minutes. Defaults to 60 minutes.}

\item{Rm}{Rate of mutation in bacterial population. Defaults to 0.001.}

\item{Ng}{Number of genes per section. One-dimensional array of length equal to number of sections. Defaults to c(6, 10, 12).}

\item{Nl}{Number of loci per gene per section. One-dimensional array of length equal to number of sections. Defaults to c(1, 6, 10).}

\item{Psize}{Starting popultaion size. Defaults to 300.}

\item{startingFitness}{Starting population fitness. Defaults to 0.51.}

\item{thr}{Bacteria with an overall fitness below this specified threshold do not reproduce to the next generation. Defaults to 0.51.}

\item{nDays}{Number of days that the bacterial population will be treated with ampicillin. Defaults to 15.}

\item{maxPsize}{The maximum population size that the bacterial population will reach within a day. Defaults to 2000.}
}
\description{
TO BE ADDED LATER
}
\author{
Ahmad Abdel-Azim \email{agabdel.azim@gmail.com}

Salma Abdel-Azim \email{salma.abdelazim@gmail.com}

Gamal Abdel-Azim \email{gamal.azim@gmail.com}
}
\keyword{SAR}
\keyword{antibiotic}
\keyword{resistance}
\keyword{simulation}
