% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_exec_sadf.R
\name{sar_exec_sadf}
\alias{sar_exec_sadf}
\title{sar_exec_sadf}
\usage{
sar_exec_sadf(statistic = c("cpu", "ram", "network", "io", "load"),
  format = c("df", "csv", "json", "xml"), server = NULL)
}
\arguments{
\item{statistic}{Statistics to retreive.}

\item{format}{Only first item will be used. The format of data to produce.}

\item{server}{list of servers to ssh into and retrieve statistics}
}
\description{
sar_exec_sadf
}
\examples{
\dontrun{
sar_exec_sadf()
sar_exec_sadf(server = "localhost")
sar_exec_sadf(format = "json")
}

}
