% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_normalize_dat.R
\name{sar_normalize_dat}
\alias{sar_normalize_dat}
\title{sar_normalize_dat}
\usage{
sar_normalize_dat(dat, node)
}
\arguments{
\item{dat}{single data item from parsed json}

\item{node}{node name}

\item{ts}{timestamp}

\item{tspan}{timespan}
}
\description{
sar_normalize_dat
}
