% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_normalize_json.R
\name{sar_normalize_json}
\alias{sar_normalize_json}
\title{sar_normalize_json}
\usage{
sar_normalize_json(json_parsed)
}
\arguments{
\item{json_parsed}{json parsed to list via jsonlite::fromJSON()}
}
\description{
sar_normalize_json
}
