% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_stat.R
\name{sar_stat}
\alias{sar_stat}
\title{sar_stat}
\usage{
sar_stat(statistic = c("cpu", "ram", "network", "io", "load"),
  as = c("df", "csv", "json", "xml"), server = NULL)
}
\arguments{
\item{statistic}{list of categories to report statistics on}

\item{as}{format to return}

\item{server}{optional server to get statistics from - if set, function
will try to ssh into server and retrieve statitics otherwise it will
retrieve statistics from local machine}
}
\description{
sar_stat
}
