% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFitness.R
\name{setFitness}
\alias{setFitness}
\title{Set Fitness for a Bacterial Population}
\usage{
setFitness(bCell, Nl, Ng)
}
\arguments{
\item{bCell}{One bacterial cell in the population, represented as a row in the population matrix, P.}

\item{Nl}{Number of loci per gene per section. One-dimensional array of length equal to number of sections. Defaults to c(1, 6, 10).}

\item{Ng}{Number of genes per section. One-dimensional array of length equal to number of sections. Defaults to c(6, 10, 12).}
}
\description{
TO BE ADDED LATER
}
\author{
Ahmad Abdel-Azim \email {agabdel.azim@gmail.com}

Salma Abdel-Azim \email {salma.abdelazim@gmail.com}

Gamal Abdel-Azim \email {gamal.azim@gmail.com}
}
