% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellDivide.R
\name{cellDivide}
\alias{cellDivide}
\title{Division of Bacterial Population}
\usage{
cellDivide(P, Rd, maxPsize, thr)
}
\arguments{
\item{P}{The population of bacteria arranged in a two-dimensional matrix of 0 and 1 values, where 0 represents naive loci and 1 represents mutated loci.}

\item{Rd}{ADD LATER


# A function to reproduce the population by cell fission/division
cellDivide0 <- function(P, Rd) {

  if(length(Rd) != nrow(P))
    stop("Number of elements in Rd != number of rows in P")

  n = length(Rd)
  x <- runif(min = 0, max = 1, n=n)
  y <- x <= Rd
  P <- P[y,]
  P <- rbind(P, P)
  P <- P[sample(1:nrow(P), size = nrow(P)),]
  P
}


cellDivide1 <- function(P, Rd, sel.intensity = .5) {

  if(length(Rd) != nrow(P))
    stop("Number of elements in Rd != number of rows in P")

  # Use order()
  x <- order(Rd)
  at = round((1-sel.intensity)*length(Rd)) + 1
  y <- Rd >= Rd[x[at]]

  P <- P[y,]
  P <- rbind(P, P)
  P <- P[sample(1:nrow(P), size = nrow(P)),]
  P
}}

\item{maxPsize}{The maximum population size that the bacterial population will reach within a day. Defaults to 2000.}

\item{thr}{Bacteria with an overall fitness below this specified threshold do not reproduce to the next generation. Defaults to 0.51.}
}
\description{
Functions to manipulate individual operations of division within a cell population.
}
\author{
Ahmad Abdel-Azim \email {agabdel.azim@gmail.com}

Salma Abdel-Azim \email {salma.abdelazim@gmail.com}

Gamal Abdel-Azim \email {gamal.azim@gmail.com}
}
