% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSectionResistance.R
\name{getSectionResistance}
\alias{getSectionResistance}
\title{Measure Resistance of Each Section of a Bacterial Genome}
\usage{
getSectionResistance(gtype, Nl, Ng)
}
\arguments{
\item{gtype}{ADD LATER}

\item{Nl}{Number of loci per gene per section. One-dimensional array of length equal to number of sections. Defaults to c(1, 6, 10).}

\item{Ng}{Number of genes per section. One-dimensional array of length equal to number of sections. Defaults to c(6, 10, 12).}
}
\description{
TO BE ADDED LATER
}
\examples{
sectionResist <- t(apply(P, 1, function(x) getSectionResistance(x, Nl = c(1, 2, 3), Ng = c(10, 10, 10))))
sectionResist

}
\author{
Ahmad Abdel-Azim \email{agabdel.azim@gmail.com}

Salma Abdel-Azim \email{salma.abdelazim@gmail.com}

Gamal Abdel-Azim \email{gamal.azim@gmail.com}
}
\keyword{SAR}
\keyword{antibiotic}
\keyword{resistance}
\keyword{simulation}
