% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_data.R
\docType{data}
\name{ms_catalog}
\alias{ms_catalog}
\title{Sample catalog dataset}
\format{
An object of class \code{data.frame} with 101 rows and 3 columns.
}
\source{
Microsoft.
}
\usage{
ms_catalog
}
\description{
A data frame containing a sample of item descriptions from the Microsoft online store. The variables are:
\itemize{
\item \code{item}: The item ID, corresponding to the items in the \link{ms_usage} dataset.
\item \code{name}: A short description of the item.
\item \code{category}: The item category.
}
}
\seealso{
\link{ms_usage}
}
\keyword{datasets}
