% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_data.R
\docType{data}
\name{ms_usage}
\alias{ms_usage}
\title{Sample usage dataset}
\format{An object of class \code{data.frame} with 11838 rows and 3 columns.}
\source{
Microsoft.
}
\usage{
ms_usage
}
\description{
A data frame of anonymised transaction records from the Microsoft online store. The variables are:
\itemize{
\item \code{user}: The user ID.
\item \code{item}: The item ID, corresponding to the items in the \link{ms_catalog} dataset.
\item \code{time}: The date and time of the transaction, in POSIXct format.
}
}
\seealso{
\link{ms_catalog}
}
\keyword{datasets}
