% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_exec_sadf.R, R/sar_stat.R
\name{sar_exec_sadf}
\alias{sar_exec_sadf}
\alias{sar_stat}
\title{sar_exec_sadf}
\usage{
sar_exec_sadf(statistic = c("cpu", "ram", "network", "io", "load"),
  format = c("df", "csv", "json", "xml"), server = NULL, day = NULL,
  log_path = NULL)

sar_stat(statistic = c("cpu", "ram", "network", "io", "load"),
  as = c("df", "json", "xml"), server = NULL, day = NULL,
  log_path = NULL)
}
\arguments{
\item{statistic}{Statistics to retreive.}

\item{format}{Only first item will be used. The format of data to produce.}

\item{server}{optional server to get statistics from - if set, function
will try to ssh into server and retrieve statitics otherwise it will
retrieve statistics from local machine}

\item{day}{optional; either a positive number indicating the calender day or a negative number indicating the number of days before today}

\item{log_path}{optional; a path to look for dat files}

\item{as}{format to return}
}
\description{
sar_exec_sadf

sar_stat
}
\examples{
\dontrun{
sar_exec_sadf()
sar_exec_sadf(server = "localhost")
sar_exec_sadf(format = "json")
}


\dontrun{
 sar_stat()
 sar_stat(as = "json")
 sar_stat(server = "localhost")
}

}
