% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarioSimulate.R
\name{scenarioSimulate}
\alias{scenarioSimulate}
\title{Simulate an Entire Genome Scenario}
\usage{
scenarioSimulate(Psize, Nl, Ng, runs, nDays, gen.interval, Rm, startingFitness, thr, maxPsize)
}
\arguments{
\item{Psize}{Starting popultaion size. Defaults to 300.}

\item{Nl}{Number of loci per gene per section. One-dimensional array of length equal to number of sections. Defaults to c(1, 6, 10).}

\item{Ng}{Number of genes per section. One-dimensional array of length equal to number of sections. Defaults to c(6, 10, 12).}

\item{runs}{ADD LATER}

\item{nDays}{ADD LATER}

\item{gen.interval}{ADD LATER}

\item{Rm}{ADD LATER}

\item{startingFitness}{Starting population fitness, represnted by the proportion of mutated bacterial cells in the population. Defaults to 0.51.}

\item{thr}{ADD LATER}

\item{maxPsize}{ADD LATER}
}
\value{
ADD LATER
}
\description{
ADD LATER
}
\details{
ADD LATER
}
\examples{
x <- scenarioSimulate()


}
\author{
Ahmad Abdel-Azim \email {agabdel.azim@gmail.com}

Salma Abdel-Azim \email {salma.abdelazim@gmail.com}

Gamal Abdel-Azim \email {gamal.azim@gmail.com}
}
\keyword{SAR}
\keyword{antibiotic}
\keyword{resistance}
\keyword{simulation}
